# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice for User Management",
        "title": "User Service API",
        "contact": {},
        "version": "1.4"
    },
    "basePath": "/api/service",
    "paths": {
        "/api/service/user": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Inserts a user",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.User"
                        }
                    }
                }
            }
        },
        "/api/service/user/{id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Updates a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Deletes a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/service/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Select users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Linus-Regander_Go-Microservice_internal_api_model_user.Pagination": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "github_com_Linus-Regander_Go-Microservice_internal_api_model_user.Role": {
            "type": "string",
            "enum": [
                "admin",
                "intern"
            ],
            "x-enum-varnames": [
                "Admin",
                "Intern"
            ]
        },
        "github_com_Linus-Regander_Go-Microservice_internal_api_model_user.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_Linus-Regander_Go-Microservice_internal_api_model_user.UserRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_Linus-Regander_Go-Microservice_internal_api_model_user.UserResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.Pagination"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Linus-Regander_Go-Microservice_internal_api_model_user.User"
                    }
                }
            }
        }
    }
}